Initial refactor of codebase structure to organize traits and implementations
