+++ Cargo.toml
@@ -24,6 +24,7 @@ serde_json = { version = "1.0", features = ["raw_value"] }
 rand_distr = "0.4.3"
 
 [features]
+default = [ "db_dependent" ]
 db_dependent = []
 
 [[bench]]
