+++ src/graph_store/mod.rs
@@ -1,0 +1,8 @@
+pub mod graph_mem;
+pub use graph_mem::GraphMem;
+
+// Postgres database graph store implementation
+#[cfg(feature = "db_dependent")]
+pub mod graph_pg;
+#[cfg(feature = "db_dependent")]
+pub use graph_pg::{test_utils::TestGraphPg, GraphPg};
